requirements.txt
google-generativeai
streamlit
faster-whisper
edge-tts
sounddevice
numpy
pyaudio
python-dotenv
watchdog
googlesearch-python
ddgs
pygame
chromadb
fastapi
uvicorn[standard]
websockets
pydantic
requests
